fix(generator): generator Path and Output pick enhancement

+++ __examples__/client_demo/types.go
@@ -137,7 +137,7 @@ func (v DemoProtocol) Value() (database_sql_driver.Value, error) {
 	if o, ok := (interface{})(v).(github_com_go_courier_enumeration.EnumDriverValueOffset); ok {
 		offset = o.Offset()
 	}
-	return int(v) + offset, nil
+	return int64(v) + int64(offset), nil
 }
 
 func (v *DemoProtocol) Scan(src interface{}) error {
